+++ src/apps/sequencer/ui/pages/GeneratorPage.h
@@ -46,6 +46,7 @@ private:
     void drawEuclideanGenerator(Canvas &canvas, const EuclideanGenerator &generator) const;
     void drawRandomGenerator(Canvas &canvas, const RandomGenerator &generator) const;
     void drawAcidGenerator(Canvas &canvas, const AcidGenerator &generator) const;
+    int contextItemCount() const;
 
     Generator *_generator;